Bonus: add identity api authorization

+++ PizzaStore/Program.cs
@@ -1,7 +1,17 @@
 using Microsoft.OpenApi.Models;
 using PizzaStore.DB;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Mvc;
+
 
 var builder = WebApplication.CreateBuilder(args);
+builder.Services.AddIdentityApiEndpoints<IdentityUser>()
+    .AddEntityFrameworkStores<ApplicationDbContext>();
+builder.Services.AddAuthorization();
+builder.Services.AddDbContext<ApplicationDbContext>(
+    options => options.UseInMemoryDatabase("AppDb"));
     
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen(c =>
@@ -10,6 +20,8 @@ builder.Services.AddSwaggerGen(c =>
 });
     
 var app = builder.Build();
+
+app.MapIdentityApi<IdentityUser>();
     
 if (app.Environment.IsDevelopment())
 {
@@ -24,8 +36,20 @@ app.MapGet("/", () => "Hello World!");
 
 app.MapGet("/pizzas/{id}", (int id) => PizzaDB.GetPizza(id));
 app.MapGet("/pizzas", () => PizzaDB.GetPizzas());
-app.MapPost("/pizzas", (Pizza pizza) => PizzaDB.CreatePizza(pizza));
-app.MapPut("/pizzas", (Pizza pizza) => PizzaDB.UpdatePizza(pizza));
-app.MapDelete("/pizzas/{id}", (int id) => PizzaDB.RemovePizza(id));
+app.MapPost("/pizzas", (Pizza pizza) => PizzaDB.CreatePizza(pizza)).RequireAuthorization();
+app.MapPut("/pizzas", (Pizza pizza) => PizzaDB.UpdatePizza(pizza)).RequireAuthorization();
+app.MapDelete("/pizzas/{id}", (int id) => PizzaDB.RemovePizza(id)).RequireAuthorization();
+
+app.MapPost("/logout", async (SignInManager<IdentityUser> signInManager,
+    [FromBody] object empty) =>
+{
+    if (empty != null)
+    {
+        await signInManager.SignOutAsync();
+        return Results.Ok();
+    }
+    return Results.Unauthorized();
+})
+.RequireAuthorization();
     
 app.Run();
